% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team10.R
\docType{package}
\name{MapDfGenerator-package}
\alias{MapDfGenerator}
\alias{MapDfGenerator-package}
\title{MapDfGenerator: MapDfGenerator}
\description{
Generate a dataframe containing location information.
}
\details{
Four functions for geographic info
}
\references{
input data is available here ""
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zeruiz/MapDfGenerator}
  \item Report bugs at \url{https://github.com/zeruiz/MapDfGenerator/issues}
}

}
\author{
\strong{Maintainer}: Zerui Zhang \email{zeruiz@iastate.edu}

Authors:
\itemize{
  \item Yudi Zhang \email{yudiz@iastate.edu}
  \item Atousa Zarindast \email{atousazarindast@gmail.com}
  \item Ying Zheng \email{zhy@iastate.edu}
}

}
\keyword{internal}
