% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team10.R
\name{shpBigToSmall}
\alias{shpBigToSmall}
\title{geographic information}
\usage{
shpBigToSmall(path, tolerance)
}
\arguments{
\item{path}{input file path usually as data/...}

\item{tolerance}{the amount of which data is getting small}
}
\value{
a data frame consists of the geographic information of the polygons and the additional information (such as name of the country, name of the territory/state, ...)
}
\description{
geographic information
}
\examples{
shpBigToSmall("./data/gadm36_AUS_shp/gadm36_AUS_1.shp", 0.1)
}
