% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team11.R
\name{team_11}
\alias{team_11}
\title{This function extracts a data from a shapefile with 2 levels in spf$geometry, and include the other information.}
\usage{
team_11(file, tolerance = 0.1)
}
\arguments{
\item{file}{Path to your shapefile.}

\item{tolerance}{The tolerance value in the metric of the input object, default is 0.1.}
}
\value{
A dataframe extracted from a shapefile with group, longitude, latitude and order and additional information.
}
\description{
This function extracts a data from a shapefile with 2 levels in spf$geometry, and include the other information.
}
