% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team12.R
\name{team_12}
\alias{team_12}
\title{Map dataframe generator form Team 12.}
\usage{
team_12(file, tolerance)
}
\arguments{
\item{file}{URL to local path to the shape data.}

\item{tolerance}{A number.}
}
\value{
The dataframe of longitude, latitude, group and order.
}
\description{
Map dataframe generator form Team 12.
}
\examples{
team_12("./data/gadm36_AUS_shp/gadm36_AUS_1.shp", 0.1)
}
