% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team5.R
\name{team_5}
\alias{team_5}
\title{The team_5 function utilize the map-dataframe-generator form team 5.}
\usage{
team_5(file, tolerance)
}
\arguments{
\item{file}{URL of local path to the shape data.}

\item{tolerance}{The tolerance value, default is 0.1.}
}
\value{
a dataframe of longitude, latitude, group, order and addtional information.
}
\description{
The team_5 function utilize the map-dataframe-generator form team 5.
}
\examples{
team_5("./data/gadm36_AUS_shp/gadm36_AUS_1.shp", 0.1)

}
