% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team12.R
\name{team_12}
\alias{team_12}
\title{This function helps get the polygon and geographic information from a shapefile.}
\usage{
team_12(file, tolerance)
}
\arguments{
\item{file}{URL to local path to the shape data.}

\item{tolerance}{A number.}
}
\value{
A dataframe, containing polygon and geographic information like longitude, latitude, group and order.
}
\description{
Extract polygon and geographic information from a shapefile. The main function is built on that from team 12's work from Lab 2
}
\examples{
filename <- system.file("extdata", "gadm36_AUS_1.shp", package = "MapDfGenerator")
dat <- team_12(filename, tolerance = 0.01)
}
